Add @zulip/shared as a dependency.
This whole zulip-markdown-parser package is an old fork of the
Markdown implementation in the Zulip web frontend.

We've begun making parts of that implementation -- the current
version of it, with all the new features and bugfixes of the last
several years -- available in the @zulip/shared package.  Adding
a dependency on that is the first step here toward replacing this
fork with that current version, piece by piece.