Unfork one module, fenced_code.
This is the first step of zulip/zulip-mobile#4242: we're taking one
module of this Markdown implementation which was forked in 2017 from
the webapp, and unforking it by switching to the current version,
provided in @zulip/shared.  We'll go on to unfork other modules one by
one until this package is a trivial wrapper for part of @zulip/shared,
and then its user (the Zulip mobile app) can switch to using the
latter directly.

Unforking fenced_code requires two updates to the rest of the code:

 * Following zulip/zulip@8e356368f, the fenced_code module no longer
   exports a function set_escape_func, and its call site can simply
   disappear.  The new version of fenced_code.js always escapes.

   (That commit made a number of other changes in other parts of the
   Markdown implementation; we'll get the benefit of those when we
   start sharing those parts too.)

 * Paralleling zulip/zulip@8e9317582, we update some test cases.

I inspected the changes within the fork, like so:

  $ git checkout 408084bf8  # this repo's initial commit
  $ git --git-dir=../zulip/.git diff {4f4d284:static/js/,}fenced_code.js
  $ git log --stat -p ..master fenced_code.js

and there's nothing that isn't taken care of upstream; it just adds
some `require` calls and converts some uses of underscore to lodash.

I also inspected the changes upstream:

  # in a zulip/zulip worktree
  $ git log --stat -p --follow 4f4d284.. static/shared/js/fenced_code.js

and there's nothing else there that looks like it would depend on
changes elsewhere.