Add and configure babel-jest as a dev dependency.
Without this, if we start using something from @zulip/shared and
then try to run Jest, it fails with a SyntaxError.

The fix is to tell Jest to have Babel compile the file first.
We can do this with the steps described here (on an issue
reporting the same error message I saw before making this fix):
  https://github.com/facebook/jest/issues/9395#issuecomment-583799300